Split formData handler to GET and POST files
